修改：似乎ringId是支持改成long的

+++ src/Application.Core/Game/Players/Wedding.cs
@@ -57,7 +57,7 @@ namespace Application.Core.Game.Players
             return getEventInstance() as Marriage;
         }
 
-        public Ring? getRingById(int id)
+        public Ring? getRingById(long id)
         {
             foreach (Ring ring in getCrushRings())
             {
